requirements.txt
gradio_imageslider
gradio==4.14.0
torch
torchvision
opencv-python
huggingface_hub
ultralytics
pillow
imutils
timm
einops
openvino
hydra-core
iopath
tqdm
numpy
matplotlib
eva-decord
scikit-image
scipy
wandb
transformers
diffusers
accelerate
huggingface_hub[hf-xet]
torchdiffeq>=0.2.3